+++ src/main/java/org/sopt/utils/IdGeneratorUtil.java
@@ -2,8 +2,8 @@ package org.sopt.utils;
 
 public class IdGeneratorUtil {
 
-    private static int postId = 1;
-    public int generateId(){
+    private static Long postId = 1L;
+    public Long generateId(){
         return postId++;
     }
 }
